CT-2238 Refactor state transitions into offender sar controller
